Move machine remediation logic from a separate contoller into actuator
